+++ src/ff7_opengl.cpp
@@ -362,10 +362,12 @@ void ff7_init_hooks(struct game_obj *_game_object)
 	if (game_lighting != GAME_LIGHTING_ORIGINAL)
 	{
 		// Disables unnecesary lighting in Chocobos applied throught the KAWAI op
-		memset_code(ff7_externals.field_apply_kawai_op_64A070 + 0x864, 0x90, 5);	
+		memset_code(ff7_externals.field_apply_kawai_op_64A070 + 0x864, 0x90, 5);
 		memset_code(ff7_externals.field_apply_kawai_op_64A070 + 0x2E4, 0x90, 5);
 		memset_code(ff7_externals.field_apply_kawai_op_64A070 + 0x3A3, 0x90, 5);
 		memset_code(ff7_externals.field_apply_kawai_op_64A070 + 0x23C, 0x90, 5);
+		// Disables unnecessary lighting in temple of the ancients rolling rocks
+		replace_function(ff7_externals.sub_64EC60, noop);
 	}
 
 	//#############################################